Updated vimrc(hlsearch) and zshrc (search-begin)
On vimrc, turned off search highlight and mouse mode, and added horizontal line
highlighting with gray background.
On zshrc, now when using up/down arrow to search commands starting with typed
in characters, the cursor jumps to the end of the command, instead of staying
at the previsouly stopped location.