added latex-suite support in vimrc & deleted dev/ in pythonpath
~/python/dev is no longer in the PYTHONPATH for python module search. This is
to avoid conflicts where modules exist in both ~/python and ~/python/dev, such
as on servers. Instead, we put a __init__.py inside the ~/python/dev
dicrectory, so that it itself is a module now. We now import anything from the
dev directory as for example  import dev.pytomo3d.tomo.